backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
django-redis==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1 
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
diff-match-patch==20241021
Django==5.1.3
django-allauth==65.3.1
django-filter==24.3
django-import-export==4.3.3
django-simple-history==3.7.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.1
tablib==3.7.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
